Move ParameterShape and ParameterInfo out of ov_interface
